add manage user access to apis getAllNotesByListId  update notes and delete notes

+++ api/src/note/note.controller.ts
@@ -5,7 +5,7 @@ import { GetUser } from 'src/auth/GetUser';
 import { JwtGuard } from 'src/auth/guards/jwt.guard';
 import { InvitationService } from 'src/invitation/invitation.service';
 import { ListService } from 'src/list/list.service';
-import { User } from 'src/user/user.schema';
+import { User, UserDocument } from 'src/user/user.schema';
 import { UserService } from 'src/user/user.service';
 import { NoteDTO } from './dtos/new-note.dto';
 import { UpdateNoteDTO } from './dtos/update-note.dto';
@@ -23,25 +23,30 @@ export class NoteController {
     private invitationService:InvitationService,
     ) {}
 
+   async userListInfo(user: UserDocument,listId:string){
+    const findList = await this.listService.find(listId);
+    let isCreator = false;
+    let permission = "";
+    if(user._id.toString() === findList.creator['_id'].toString()){
+       isCreator = true;
+    }else{     
+     const invitaion = await this.invitationService.findInvitaion(listId,user._id.toString());
+     if(invitaion.length !== 0)
+       permission = invitaion[0]['permission'];
+    }
+    const userListInfo = {isCreator:isCreator,permission:permission};
+    return userListInfo;
+   }
+
   @ApiBearerAuth()
   @UseGuards(JwtGuard)
   @ApiUnauthorizedResponse()
   @ApiBody({ type: NoteDTO })
   @Post('create')
   async create(@Body() note: NoteDTO,@GetUser() currentUser): Promise<NoteDocument | null> {
-    
-    const user = await this.userService.findByEmail(currentUser.email);
-    const findList = await this.listService.find(note.listId);
-    let isCreator = false;
-    let permission = "";
-    if(user._id.toString() === findList.creator['_id'].toString()){
-       isCreator = true;
-    }else{     
-     const invitaion = await this.invitationService.findInvitaion(note.listId,user._id.toString());
-     if(invitaion.length !== 0)
-       permission = invitaion[0]['permission'];
-    }
-    const ability = this.abilityFactory.defineAbility(user,isCreator,permission)
+     const user = await this.userService.findByEmail(currentUser.email);
+     let userListInfo =await this.userListInfo(user, note.listId)
+    const ability = this.abilityFactory.defineAbility(user,userListInfo.isCreator,userListInfo.permission);
     const isAllowed = ability.can(Action.Create, Note);
     if(!isAllowed)
      throw new ForbiddenException('only creator and user with readwrite permission can add notes!!');
@@ -58,9 +63,15 @@ export class NoteController {
     name: 'listId'
   })
   @Get(':listId')
-  findAllNotes(
-    @Param('listId') listId: string,
+  async findAllNotes(@Param('listId') listId: string,@GetUser() currentUser
   ): Promise<NoteDocument[]> {   
+    const user = await this.userService.findByEmail(currentUser.email);
+    let userListInfo =await this.userListInfo(user, listId)
+    const ability = this.abilityFactory.defineAbility(user,userListInfo.isCreator,userListInfo.permission);
+    const isAllowed = ability.can(Action.Read, Note);
+    if(!isAllowed)
+     throw new ForbiddenException('user is not invited to this list!!');
+
     return this.noteService.findAllByListId(listId);
   }
 
@@ -73,10 +84,18 @@ export class NoteController {
     name: 'id'
   })
   @Patch(':id')
-  updateNote(
+  async updateNote(
     @Param('id') id: string,
-    @Body() updateNoteDTO : UpdateNoteDTO
+    @Body() updateNoteDTO : UpdateNoteDTO,@GetUser() currentUser
   ): Promise<NoteDocument> {   
+    const list = await this.noteService.find(id);
+    const user = await this.userService.findByEmail(currentUser.email);
+    let userListInfo =await this.userListInfo(user, list._id.toString())
+    const ability = this.abilityFactory.defineAbility(user,userListInfo.isCreator,userListInfo.permission);
+    const isAllowed = ability.can(Action.Update, Note);
+    if(!isAllowed)
+     throw new ForbiddenException('only creator and user with readwrite permission can update notes!!');
+
     return this.noteService.update(id, updateNoteDTO);
   }
 
@@ -88,10 +107,14 @@ export class NoteController {
     name: 'id'
   })
   @Delete(':id')
-  deleteNote(@Param('id') id: string) {
+  async deleteNote(@Param('id') id: string,@GetUser() currentUser) {
+    const list = await this.noteService.find(id);
+    const user = await this.userService.findByEmail(currentUser.email);
+    let userListInfo =await this.userListInfo(user, list._id.toString())
+    const ability = this.abilityFactory.defineAbility(user,userListInfo.isCreator,userListInfo.permission);
+    const isAllowed = ability.can(Action.Update, Note);
+    if(!isAllowed)
+     throw new ForbiddenException('only creator and user with readwrite permission can delete notes!!');
     return this.noteService.delete(id);
   }
-
-   
-
 }
